Printed-Message
In this code, an outer do-while loop is added, which will run until the user presses the 'S' key to exit. Inside the outer loop, the inner do-while loop is maintained, which prompts the user to enter a valid number. After displaying the corresponding message for the entered number, the user is prompted to press 'S' to exit or any other key to continue. If the user presses 'S', the exit variable is set to true, and the outer loop is exited. If the user presses any other key, the screen is cleared, and a new number is requested.